+++ R/symbol.R
@@ -36,7 +36,7 @@ construct_symbol_from_pyobj <- function(pyobj) {
 #' if (has_sympy()) {
 #'    x <- symbol('x')
 #'    (1+1)*x^2
-#'    limf(sin(x)/x, "x", 0)
+#'    lim(sin(x)/x, "x", 0)
 #' }
 #' 
 #' @return A `caracas_symbol`
@@ -136,7 +136,7 @@ is_atomic <- function(x) {
 #' @examples 
 #' if (has_sympy()) {
 #'    x <- symbol('x')
-#'    res <- limf(sin(x)/x, "x", 0, doit = FALSE)
+#'    res <- lim(sin(x)/x, "x", 0, doit = FALSE)
 #'    res 
 #'    doit(res)
 #' }