+++ tests/testthat/test-calculus.R
@@ -54,73 +54,73 @@ test_that("limit", {
   
   x <- symbol("x")
   
-  expect_equal(as.character(limf(sin(x)/x, "x", 0)), "1")
-  expect_equal(as.character(limf(sin(x)/x, x, 0)), "1")
-  expect_equal(as_expr(limf(sin(x)/x, "x", 0)), 1)
-  expect_equal(as_expr(limf(sin(x)/x, x, 0)), 1)
+  expect_equal(as.character(lim(sin(x)/x, "x", 0)), "1")
+  expect_equal(as.character(lim(sin(x)/x, x, 0)), "1")
+  expect_equal(as_expr(lim(sin(x)/x, "x", 0)), 1)
+  expect_equal(as_expr(lim(sin(x)/x, x, 0)), 1)
   
-  expect_equal(as.character(limf(1/x, "x", 0, dir = '+')), "Inf")
-  expect_equal(as.character(limf(1/x, x, 0, dir = '+')), "Inf")
-  expect_equal(as_expr(limf(1/x, "x", 0, dir = '+')), Inf)
-  expect_equal(as_expr(limf(1/x, x, 0, dir = '+')), Inf)
+  expect_equal(as.character(lim(1/x, "x", 0, dir = '+')), "Inf")
+  expect_equal(as.character(lim(1/x, x, 0, dir = '+')), "Inf")
+  expect_equal(as_expr(lim(1/x, "x", 0, dir = '+')), Inf)
+  expect_equal(as_expr(lim(1/x, x, 0, dir = '+')), Inf)
   
-  expect_equal(as.character(limf(1/x, "x", 0, dir = '-')), "-Inf")
-  expect_equal(as.character(limf(1/x, x, 0, dir = '-')), "-Inf")
-  expect_equal(as_expr(limf(1/x, "x", 0, dir = '-')), -Inf)
-  expect_equal(as_expr(limf(1/x, x, 0, dir = '-')), -Inf)
+  expect_equal(as.character(lim(1/x, "x", 0, dir = '-')), "-Inf")
+  expect_equal(as.character(lim(1/x, x, 0, dir = '-')), "-Inf")
+  expect_equal(as_expr(lim(1/x, "x", 0, dir = '-')), -Inf)
+  expect_equal(as_expr(lim(1/x, x, 0, dir = '-')), -Inf)
   
-  expect_equal(as.character(limf((1 + 1/x)^x, x, Inf)), "exp(1)")
-  expect_equal(as_expr(limf((1 + 1/x)^x, x, Inf)), exp(1))
-  expect_equal(as_expr(-limf((1 + 1/x)^x, x, Inf)), -exp(1))
-  expect_equal(as_expr(2*limf((1 + 1/x)^x, x, Inf)), 2*exp(1))
-  expect_equal(as_expr(2^limf((1 + 1/x)^x, x, Inf)), 2^exp(1))
-  expect_equal(as_expr(limf((1 + 1/x)^x, x, Inf)^3), exp(1)^3)
-  expect_equal(as_expr(2*limf((1 + 1/x)^x, x, Inf)^3), 2*exp(1)^3)
+  expect_equal(as.character(lim((1 + 1/x)^x, x, Inf)), "exp(1)")
+  expect_equal(as_expr(lim((1 + 1/x)^x, x, Inf)), exp(1))
+  expect_equal(as_expr(-lim((1 + 1/x)^x, x, Inf)), -exp(1))
+  expect_equal(as_expr(2*lim((1 + 1/x)^x, x, Inf)), 2*exp(1))
+  expect_equal(as_expr(2^lim((1 + 1/x)^x, x, Inf)), 2^exp(1))
+  expect_equal(as_expr(lim((1 + 1/x)^x, x, Inf)^3), exp(1)^3)
+  expect_equal(as_expr(2*lim((1 + 1/x)^x, x, Inf)^3), 2*exp(1)^3)
   
   a <- symbol("a")
-  expect_equal(as.character(a*limf(sin(x)/x, "x", 0)), "a")
-  expect_equal(as.character(a*limf(sin(x)/x, x, 0)), "a")
+  expect_equal(as.character(a*lim(sin(x)/x, "x", 0)), "a")
+  expect_equal(as.character(a*lim(sin(x)/x, x, 0)), "a")
 })
 
-test_that("sumf", {
+test_that("sum_", {
   skip_if_no_sympy()
   
   x <- symbol("x")
-  s <- sumf(1/x, "x", 1, 10)
+  s <- sum_(1/x, "x", 1, 10)
   
   expect_equal(as.character(s), "7381/2520")
   expect_equal(as_expr(s), sum(1/(1:10)))
                
   n <- symbol("n")
-  s <- simplify(sumf(x, x, 1, n))
+  s <- simplify(sum_(x, x, 1, n))
   expect_equal(as.character(s), "n*(n + 1)/2")
 })
 
-test_that("prodf", {
+test_that("prod_", {
   skip_if_no_sympy()
   
   x <- symbol("x")
-  p <- prodf(1/x, "x", 1, 10)
+  p <- prod_(1/x, "x", 1, 10)
   
   expect_equal(as.character(p), "1/3628800")
   expect_equal(as_expr(p), prod(1/(1:10)))
 })
 
-test_that("intf", {
+test_that("int", {
   skip_if_no_sympy()
   
   x <- symbol("x")
   
   # Limits
-  expect_equal(as.character(intf(1/x, "x", 1, 10)), "log(10)")
-  expect_equal(as.character(intf(1/x, x, 1, 10)), "log(10)")
+  expect_equal(as.character(int(1/x, "x", 1, 10)), "log(10)")
+  expect_equal(as.character(int(1/x, x, 1, 10)), "log(10)")
   
-  i1 <- intf(1/x, "x", 1, 10, doit = FALSE)
+  i1 <- int(1/x, "x", 1, 10, doit = FALSE)
   expect_equal(as.character(i1), "Integral(1/x, (x, 1, 10))")
   expect_equal(tex(i1), "\\int\\limits_{1}^{10} \\frac{1}{x}\\, dx")
   expect_equal(as.character(doit(i1)), "log(10)")
   
-  i1 <- intf(1/x, x, 1, 10, doit = FALSE)
+  i1 <- int(1/x, x, 1, 10, doit = FALSE)
   expect_equal(as.character(i1), "Integral(1/x, (x, 1, 10))")
   expect_equal(tex(i1), "\\int\\limits_{1}^{10} \\frac{1}{x}\\, dx")
   expect_equal(as.character(doit(i1)), "log(10)")
@@ -128,17 +128,32 @@ test_that("intf", {
   
   ## No limits
   z <- symbol('z')
-  expect_equal(as.character(intf(1/z, "z")), "log(z)")
-  expect_equal(as.character(intf(1/z, z)), "log(z)")
+  expect_equal(as.character(int(1/z, "z")), "log(z)")
+  expect_equal(as.character(int(1/z, z)), "log(z)")
   
-  i1 <- intf(1/x, "x", doit = FALSE)
+  i1 <- int(1/x, "x", doit = FALSE)
   expect_equal(as.character(i1), "Integral(1/x, x)")
   expect_equal(tex(i1), "\\int \\frac{1}{x}\\, dx")
   expect_equal(as.character(doit(i1)), "log(x)")
   
-  i1 <- intf(1/x, x, doit = FALSE)
+  i1 <- int(1/x, x, doit = FALSE)
   expect_equal(as.character(i1), "Integral(1/x, x)")
   expect_equal(tex(i1), "\\int \\frac{1}{x}\\, dx")
   expect_equal(as.character(doit(i1)), "log(x)")
+})
+
+
+
+test_that("taylor", {
+  skip_if_no_sympy()
+  
+  x <- symbol("x")
+  f <- cos(x)
+  ft_with_O <- taylor(f, x0 = 0, n = 4+1)
+  
+  expect_equal(as.character(ft_with_O), "1 - x^2/2 + x^4/24 + O(x^5.0)")
+  
+  ft <- ft_with_O %>% drop_remainder()
+  expect_equal(as.character(ft), "x^4/24 - x^2/2 + 1")
 })